NVME Initial changes

+++ kernel/include/time.h
@@ -33,7 +33,7 @@ time_t mktime(struct tm *tmbuf);
 time_t mktime_simple(const struct tm *time);
 struct tm *gmtime(const time_t *timer, struct tm *tmbuf);
 void kwait(uint64_t msToWait);
-void wait(int msToWait);
+void wait(uint64_t msToWait);
 void __attribute__((noinline))waitTicks(int TicksToWait);
 
 #endif // TIME_H
